Extract core enums and scalars to separate files
