+++ saleor/graphql/checkout/types.py
@@ -3,7 +3,7 @@ import graphene_django_optimizer as gql_optimizer
 
 from ...checkout import models
 from ...core.utils.taxes import get_taxes_for_address
-from ..core.types.common import CountableDjangoObjectType
+from ..core.connection import CountableDjangoObjectType
 from ..core.types.money import TaxedMoney
 from ..order.resolvers import resolve_shipping_methods
 from ..shipping.types import ShippingMethod
