+++ saleor/graphql/core/connection.py
@@ -1,5 +1,6 @@
 import graphene
 from graphene import Field, List, NonNull, ObjectType, String
+from graphene_django import DjangoObjectType
 from graphene.relay.connection import Connection
 
 
@@ -41,3 +42,17 @@ class CountableConnection(NonNullConnection):
     @staticmethod
     def resolve_total_count(root, info, *args, **kwargs):
         return root.length
+
+
+class CountableDjangoObjectType(DjangoObjectType):
+    class Meta:
+        abstract = True
+
+    @classmethod
+    def __init_subclass_with_meta__(cls, *args, **kwargs):
+        # Force it to use the countable connection
+        countable_conn = CountableConnection.create_type(
+            "{}CountableConnection".format(cls.__name__),
+            node=cls)
+        super().__init_subclass_with_meta__(
+            *args, connection=countable_conn, **kwargs)
