+++ saleor/graphql/core/scalars.py
@@ -1,0 +1,27 @@
+import decimal
+import graphene
+
+
+class Decimal(graphene.Float):
+    """Custom Decimal implementation.
+
+    Returns Decimal as a float in the API,
+    parses float to the Decimal on the way back.
+    """
+
+    @staticmethod
+    def parse_literal(node):
+        try:
+            return decimal.Decimal(node.value)
+        except decimal.DecimalException:
+            return None
+
+    @staticmethod
+    def parse_value(value):
+        try:
+            # Converting the float to str before parsing it to Decimal is
+            # necessary to keep the decimal places as typed
+            value = str(value)
+            return decimal.Decimal(value)
+        except decimal.DecimalException:
+            return None
