+++ saleor/graphql/discount/types.py
@@ -5,8 +5,8 @@ import graphene_django_optimizer as gql_optimizer
 from graphene import relay
 
 from ...discount import DiscountValueType, VoucherType, models
+from ..core.connection import CountableDjangoObjectType
 from ..core.fields import PrefetchingConnectionField
-from ..core.types.common import CountableDjangoObjectType
 from ..product.types import Category, Collection, Product
