+++ saleor/graphql/product/mutations/products.py
@@ -11,8 +11,10 @@ from ....product.thumbnails import (
     create_category_background_image_thumbnails,
     create_collection_background_image_thumbnails, create_product_thumbnails)
 from ....product.utils.attributes import get_name_from_attributes
+from ...core.enums import TaxRateType
 from ...core.mutations import BaseMutation, ModelDeleteMutation, ModelMutation
-from ...core.types import Decimal, SeoInput, TaxRateType, Upload
+from ...core.scalars import Decimal
+from ...core.types import SeoInput, Upload
 from ...core.utils import clean_seo_fields
 from ...shipping.types import WeightScalar
 from ..types import Category, Collection, Product, ProductImage, ProductVariant
