+++ saleor/graphql/product/schema.py
@@ -4,7 +4,7 @@ import graphene
 from graphql_jwt.decorators import permission_required
 
 from ..core.fields import PrefetchingConnectionField
-from ..core.types import ReportingPeriod
+from ..core.enums import ReportingPeriod
 from ..descriptions import DESCRIPTIONS
 from .mutations.attributes import (
     AttributeCreate, AttributeDelete, AttributeUpdate, AttributeValueCreate,
