+++ saleor/graphql/shipping/mutations.py
@@ -5,7 +5,7 @@ import graphene
 from ...dashboard.shipping.forms import default_shipping_zone_exists
 from ...shipping import models
 from ..core.mutations import ModelDeleteMutation, ModelMutation
-from ..core.types.common import Decimal
+from ..core.scalars import Decimal
 from .types import ShippingMethodTypeEnum, ShippingZone, WeightScalar
