+++ tests/api/test_core.py
@@ -3,7 +3,7 @@ from unittest.mock import Mock
 import graphene
 from django.utils import timezone
 
-from saleor.graphql.core.types import ReportingPeriod
+from saleor.graphql.core.enums import ReportingPeriod
 from saleor.graphql.core.utils import clean_seo_fields, snake_to_camel_case
 from saleor.graphql.product import types as product_types
 from saleor.graphql.utils import get_database_id, reporting_period_to_date
