+++ tests/api/test_discount.py
@@ -2,7 +2,7 @@ import graphene
 import pytest
 
 from saleor.discount import DiscountValueType, VoucherType
-from saleor.graphql.discount.types import (
+from saleor.graphql.discount.enums import (
     DiscountValueTypeEnum, VoucherTypeEnum)
 from tests.api.utils import get_graphql_content
