+++ tests/api/test_order.py
@@ -4,7 +4,7 @@ import graphene
 import pytest
 
 from saleor.core.utils.taxes import ZERO_TAXED_MONEY
-from saleor.graphql.core.types import ReportingPeriod
+from saleor.graphql.core.enums import ReportingPeriod
 from saleor.graphql.order.mutations.orders import (
     clean_order_cancel, clean_order_capture, clean_order_mark_as_paid,
     clean_refund_payment, clean_void_payment)
