+++ tests/api/test_product.py
@@ -8,7 +8,7 @@ from django.utils.text import slugify
 from graphql_relay import to_global_id
 from prices import Money
 
-from saleor.graphql.core.types import ReportingPeriod
+from saleor.graphql.core.enums import ReportingPeriod
 from saleor.graphql.product.types import (
     StockAvailability, resolve_attribute_list)
 from saleor.product.models import (
